+++ lib/yokogaki.dart
@@ -9,6 +9,7 @@ export 'src/widgets/horizontal_text.dart';
 export 'src/models/horizontal_text_style.dart';
 export 'src/models/ruby_text.dart';
 export 'src/models/kenten.dart';
+export 'src/models/warichu.dart';
 
 // Re-export kinsoku package for convenience
 export 'package:kinsoku/kinsoku.dart';